refactor(onboarding): extract pure helpers from OnboardingChat
Behavior-preserving move (OnboardingChat.tsx 1222 -> ~1011): the pure
message/tool-output parsing logic (18 type guards + transforms:
getMessageText, isToolPart, getToolName, getToolParts, is*Payload/is*Output
guards, getInputQuery, findLastAssistantMessageId, getOnboardingErrorMessage,
artistFrom*, cleanHandle, deriveProfileBuilderState) plus the ToolPart /
ToolOutputWithAction types and the THINKING_PLACEHOLDER_ID constant move to
a new onboardingChatHelpers.ts — a pure, testable logic module (no JSX/hooks).
The JSX tool renderers and the OnboardingChat component import them back.
Bodies verbatim.
